requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.6
langchain-community==0.0.19
langchain-groq==0.0.1
weaviate-client==4.4.0
sentence-transformers==2.3.1
pypdf==4.0.0
pytesseract==0.3.10
tabula-py==2.9.0
Pillow==10.2.0
ragas==0.1.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
